Refactor: delete usless Ingredient class
- make DatabaseOperation package
